requirements.txt
PySocks
requests==2.20.0
python-telegram-bot
AoikLiveReload
sqlalchemy
mysqlclient
flask
flask-admin
flask-babelex
